mostly cleanup and more gutting of main boot file fixed bug in drawPixel where only a very small subset of colours could be drawn changed HX711 calibration to milligrams and added milligram mass selection to terminal interface added font file to repo

+++ src/boot/ksdk1.1.0/devHX711.h
@@ -68,7 +68,7 @@ void tareHX711(uint16_t num);
 
 // set mass calibration scale. Requires known calibration weight!
 // num = how many times to read the tare/offset values
-void calibrateHX711(uint16_t num, uint16_t cal_mass);
+void calibrateHX711(uint16_t num, uint32_t cal_mass);
 
 // return calibrated value
 int32_t readCalibratedHX711();
